Backup
Backing up my custom keymapping progress. No menu yet, but the code is ALL in place for custom keymapping! I mean, you could modify the savefiles with the literals that translate into SDL_Scancodes, but please, if you can do that with ease, you need to be helping me with this project :P